Make sure that the CIRCLE_NODE_TOTAL matches the expectation.
In a circumstance when the CI provider is not configured to run the same number of containers, not having this check allows certain ranges of these tests to not run.

I experienced this when running Meteor CircleCI on my own fork and only half the tests ran, but I still got green at the end since the ci.sh script didn't know the other containers were never spawned.